refactor(#60): relocate gate smokes examples/smoke_*.rb -> prep/smokes/
The smoke_* fixtures are GATES (the byte-identical suite wired into
docs/gating.md + make gate-*), not tutorials — move them to gate-land
per the demos-rework plan. Pure relocation, byte-exact:

- 20 fixtures git-mv'd; require_relative depth ../ -> ../../; in-file
  usage headers updated to the new make/run paths.
- Makefile: all smoke targets/paths now prep/smokes/...
- prep/gen_cuda_mirror.rb: MIRRORABLE entry, per-file rewrite rule
  (require depth) + derive_mirror_path regex follow the move.
- .gitignore: prep/smokes binary glob; generated projection-lens
  mirrors repathed.
- prep/README.md documents the smokes/ family; docs/{gating,authoring,
  framework,consuming-toy,cli,reference/loader-api}.md + lib comments
  + examples READMEs repathed.

Gate evidence (SPINEL_DIR=/tmp/spinel-a699cf9-serve, gx10):
verify-mirrors, gate-compute-surface(+cuda), recipe==lens byte-exact,
all 19 CPU smokes rebuilt+run, infer/train/eval/run_log/gpt2(+engine,
+cuda)/ckpt_roundtrip/lmc/full_finetune/mixed_precision/train_cuda/
moe_kquant/gate-cuda ALL PASS. poly_degrade still exits 1 on a699cf9
(pre-existing baseline artifact, unchanged).
