+++ docs/consuming-toy.md
@@ -74,7 +74,7 @@ so `bin/toy` can require it; don't use `toy.rb` for the compute surface. If you
 build only one engine and want to compile less (Spinel has no tree-shaking),
 require that engine's file directly instead of `toy/compute`. The
 `gate-compute-surface` make target proves the one-require surface co-compiles +
-runs a from-scratch training step (`examples/smoke_compute_surface`).
+runs a from-scratch training step (`prep/smokes/smoke_compute_surface`).
 
 > One exception: `toy/compute` does **not** pull `recipes/lora` — loading it
 > uncalled trips a Spinel analyzer bug that poisons `cfg` to poly program-wide
