+++ prep/gen_cuda_mirror.rb
@@ -70,7 +70,7 @@ MIRRORABLE = [
   "lib/gpt2_ffi.rb",                    # GPT-2 full-forward FFI (Phase 0.6 step 3)
   "lib/gpt2_ffi_kv.rb",                 # GPT-2 KV-cache decode (Phase 0.6 step 3)
   "examples/06_train_from_scratch.rb",  # #152: from-scratch training entry (CPU/CUDA)
-  "examples/smoke_projection_lens.rb",  # P2.6 CUDA gate: realize_for_random_init forward on GPU
+  "prep/smokes/smoke_projection_lens.rb", # P2.6 CUDA gate: realize_for_random_init forward on GPU
 ]
 
 # Per-backend codegen parameters.
@@ -366,15 +366,17 @@ def subs_for(cpu_path, backend)
       [/\bLlamaSeqEngine\b/, "LlamaSeqEngine" + suffix],
     ] + common_module_tail
 
-  when "examples/smoke_projection_lens.rb"
+  when "prep/smokes/smoke_projection_lens.rb"
     # P2.6 CUDA gate — mirror of the projection-lens smoke so the
     # realize_for_random_init forward can be parity-gated on the GPU
     # backend (CUDA self-consistency, not CPU bit-equality). Same
     # rewrites as the 06 example: require the GPU engine mirror,
     # rename the engine class, TinyNN. -> TinyNN<Suffix>. via the tail.
+    # (Smokes live two levels below repo root — prep/smokes/ — hence
+    # the ../../ require depth, unlike the examples/ entry above.)
     [
-      [/^require_relative "..\/lib\/toy\/llm\/engine\/llama_seq_engine"$/,
-       'require_relative "../lib/toy/llm/engine/llama_seq_engine_' + backend.to_s + '"'],
+      [/^require_relative "..\/..\/lib\/toy\/llm\/engine\/llama_seq_engine"$/,
+       'require_relative "../../lib/toy/llm/engine/llama_seq_engine_' + backend.to_s + '"'],
       [/\bLlamaSeqEngine\b/, "LlamaSeqEngine" + suffix],
     ] + common_module_tail
 
@@ -424,7 +426,7 @@ def derive_mirror_path(cpu_path, backend)
   # lib/toy/llm/primitives/<name>.rb → ..._<backend>.rb (P2.3 L1 primitives)
   unless cpu_path =~ /_ffi(_[a-z]+)?\.rb$/ ||
          cpu_path =~ %r{^examples/\d+_[a-z0-9_]+\.rb$} ||
-         cpu_path =~ %r{^examples/smoke_[a-z0-9_]+\.rb$} ||
+         cpu_path =~ %r{^prep/smokes/smoke_[a-z0-9_]+\.rb$} ||
          cpu_path =~ %r{^lib/toy/llm/primitives/[a-z0-9_]+\.rb$} ||
          cpu_path =~ %r{^lib/toy/llm/blocks/[a-z0-9_]+\.rb$} ||
          cpu_path =~ %r{^lib/toy/llm/archs/[a-z0-9_]+\.rb$} ||
