+++ prep/smokes/smoke_gate_b_gt_1.rb
@@ -16,8 +16,8 @@
 #
 # Pure-Ruby CPU, Spinel-compiled. NO lib/ behavior change, NO mirror regen.
 #
-#   make examples/smoke_gate_b_gt_1
-#   STEPS=5 SEED=0 ./examples/smoke_gate_b_gt_1   (run from repo root)
+#   make prep/smokes/smoke_gate_b_gt_1
+#   STEPS=5 SEED=0 ./prep/smokes/smoke_gate_b_gt_1   (run from repo root)
 #
 # Determinism: SEED=0 + weight_dtype=0 (F32, no F16/BF16 reduction-order
 # nondeterminism across the larger T*B score matrix) + fixed FIRST 2 lines
@@ -30,11 +30,11 @@
 # ONLY about batching. head_dim defaults to d_model/n_heads (16) so
 # n_heads*head_dim == d_model (no w_o divergence — that is a separate gate).
 
-require_relative "../lib/toy"
-require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/llm/engine/llama_seq_engine"
-require_relative "../lib/toy/llm/adamw"
-require_relative "../lib/toy/train/toy_drift_grad"
+require_relative "../../lib/toy"
+require_relative "../../lib/toy/models/toy_smollm2"
+require_relative "../../lib/toy/llm/engine/llama_seq_engine"
+require_relative "../../lib/toy/llm/adamw"
+require_relative "../../lib/toy/train/toy_drift_grad"
 
 STEPS    = (ENV["STEPS"] || "5").to_i
 SEED     = (ENV["SEED"]  || "0").to_i
