+++ prep/smokes/smoke_gate_llama3_tensor.rb
@@ -48,12 +48,12 @@
 # Spinel: needs compilation (ffi_lib is a Spinel intrinsic). NO Struct.new
 # (#16; cfg objects are hand-written ctors). NO lib/ change. One commit.
 #
-#   make examples/smoke_gate_llama3_tensor
-#   ./examples/smoke_gate_llama3_tensor        # run from repo root
+#   make prep/smokes/smoke_gate_llama3_tensor
+#   ./prep/smokes/smoke_gate_llama3_tensor        # run from repo root
 
-require_relative "../lib/toy"
-require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/llm/engine/llama_seq_engine"
+require_relative "../../lib/toy"
+require_relative "../../lib/toy/models/toy_smollm2"
+require_relative "../../lib/toy/llm/engine/llama_seq_engine"
 
 # llama-3.2 rope params (factor=8.0, low=1.0, high=4.0, orig_max_pos=8192,
 # base=500000) + d_head=64. T sequence positions, single batch.
